Build configuration will use corresponding build of glew32s
Previously, both DEBUG and RELEASE configurations were hard linked to utilize the RELEASE build. This has been changed to use the '$(Configuration)' variable to pack the corresponding build of glew32. In other words, the DEBUG build of the AIEFramework will use the DEBUG build of GLEW and the RELEASE build of AIE would use the RELEASE build of glew32.